Move status bar font-related code
