+++ foo_ui_columns/font_notify.cpp
@@ -1,82 +1,7 @@
 #include "stdafx.h"
 #include "ng_playlist/ng_playlist.h"
 
-#include "status_pane.h"
 #include "font_notify.h"
-#include "status_bar.h"
-
-HFONT g_status_font = nullptr;
-HFONT g_tab_font = nullptr;
-HFONT g_plist_font = nullptr;
-
-void font_cleanup()
-{
-    if (g_status_font) {
-        DeleteObject(g_status_font);
-        g_status_font = nullptr;
-    }
-    if (g_tab_font) {
-        DeleteObject(g_tab_font);
-        g_tab_font = nullptr;
-    }
-}
-
-// {B9D5EA18-5827-40be-A896-302A71BCAA9C}
-static const GUID font_client_status_guid
-    = {0xb9d5ea18, 0x5827, 0x40be, {0xa8, 0x96, 0x30, 0x2a, 0x71, 0xbc, 0xaa, 0x9c}};
-
-class font_client_status : public cui::fonts::client {
-public:
-    const GUID& get_client_guid() const override { return font_client_status_guid; }
-    void get_name(pfc::string_base& p_out) const override { p_out = "Status bar"; }
-
-    cui::fonts::font_type_t get_default_font_type() const override { return cui::fonts::font_type_labels; }
-
-    void on_font_changed() const override { on_status_font_change(); }
-};
-
-font_client_status::factory<font_client_status> g_font_client_status;
-
-void on_status_font_change()
-{
-    {
-        if (g_status_font != nullptr) {
-            if (g_status)
-                SendMessage(g_status, WM_SETFONT, (WPARAM)0, MAKELPARAM(0, 0));
-            DeleteObject(g_status_font);
-            g_status_font = nullptr;
-        }
-
-        if (g_status) {
-            g_status_font = static_api_ptr_t<cui::fonts::manager>()->get_font(font_client_status_guid);
-            SendMessage(g_status, WM_SETFONT, (WPARAM)g_status_font, MAKELPARAM(1, 0));
-            status_bar::set_part_sizes(status_bar::t_parts_all);
-            cui::main_window.resize_child_windows();
-        }
-    }
-}
-
-#if 0
-void on_tab_font_change()
-{
-    {
-        if (g_tab_font!=0)
-        {
-            if (g_tab) SendMessage(g_tab,WM_SETFONT,(WPARAM)0,MAKELPARAM(0,0));
-            DeleteObject(g_tab_font);
-            g_tab_font=0;
-        }
-        
-        if (g_tab) 
-        {
-            g_tab_font = CreateFontIndirect(&(LOGFONT)cfg_tab_font);
-            SendMessage(g_tab,WM_SETFONT,(WPARAM)g_tab_font,MAKELPARAM(1,0));
-            size_windows();
-            
-        }
-    }
-}
-#endif
 
 int CALLBACK FontSizesProc(const LOGFONT* plf, const TEXTMETRIC* ptm, DWORD FontType, LPARAM lp);
